requirements.txt
console-menu
grpcio-tools
grpcio
prettytable
psutil
pywin32
typing-extensions
